impl: cli

+++ src/main.py
@@ -1,11 +1,27 @@
-from os import system
+from os import system, getcwd, chdir
+import argparse
 
 def main():
+  root_path = f'{getcwd()}/{__file__}'.split('src')[0]
+
+  parser = argparse.ArgumentParser('Ferramenta para auxiliar na comunicação com o Raspberry')
+  parser.add_argument('--install', help='Configure the container', action='store_true')
+  parser.add_argument('--start', help='Start the server', action='store_true')
+
+  args = parser.parse_args()
+
+  print(root_path)
+
+  if (not len(args._get_args())):
+    parser.print_help()
+    exit(0)
+
+  if (args.install):
     system('sudo apt-get update -y')
     system('sudo apt-get upgrade -y')
     system('sudo apt-get autoremove -y')
 
-    system('sudo apt-get install apt-transport-https ca-certificates curl gnupg2 software-properties-common -y')
+    system('sudo apt-get screen install docker.io apt-transport-https ca-certificates curl gnupg2 software-properties-common -y')
     system('curl -fsSL https://download.docker.com/linux/debian/gpg | sudo apt-key add -')
     system('sudo apt-key fingerprint 0EBFCD88')
     system('sudo add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/debian $(lsb_release -cs) stable"')
@@ -16,6 +32,14 @@ def main():
     system('sudo chmod +x /usr/local/bin/docker-compose')
     system('sudo ln -s /usr/local/bin/docker-compose /usr/bin/docker-compose')
 
+    chdir(f'{root_path}/src')
+    system('chmod a+x ./start.sh')
+
+    print('Success configuring server')
+
+  if (args.start):
+    chdir(f'{root_path}/container')
+    system('docker-compose up -d')
 
 if __name__ == '__main__':
-    main()
+  main()
